Extract util functions from app.py to new files
Extract utility functions from `app.py` to a new file `utils.py`.

* **Add `utils.py` file:**
  - Add function `count_tokens` to count the number of tokens in a text string.
  - Add function `split_text_into_chunks` to split text into chunks of maximum token size.
  - Add function `extract_text_from_pdf` to extract text from a PDF file and return text chunks and their token counts.
  - Add function `get_summary` to get summary of text using OpenAI.
  - Add function `process_document_chunks` to process multiple chunks of a document and return their data.
  - Add function `select_relevant_document` to select the most relevant document based on the question and summaries.
  - Add function `get_answer` to get answer to question using the selected document.

* **Modify `app.py` file:**
  - Import utility functions from `utils.py`.
  - Remove function `count_tokens`.
  - Remove function `split_text_into_chunks`.
  - Remove function `extract_text_from_pdf`.
  - Remove function `get_summary`.
  - Remove function `process_document_chunks`.
  - Remove function `select_relevant_document`.
  - Remove function `get_answer`.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/melkor54248/noRAG-multiagent-doc-qna?shareId=XXXX-XXXX-XXXX-XXXX).
